cleanup now checks if the files exist before attempting to remove

+++ build/cleanup.sh
@@ -3,12 +3,12 @@
 set -e
 
 echo "Removing raw assembly"
-rm ./*.asm
-rm ./*.wasm #in case someone modifies the source code to produce .wasm files
-rm ./*.nasm #in case someone modifies the source code to produce .nasm files
-rm ./*.fasm #in case someone modifies the source code to produce .fasm files
+[ -f "$./*.asm" ] && rm ./*.asm
+[ -f "$./*.wasm" ] && rm ./*.wasm #in case someone modifies the source code to produce .wasm files
+[ -f "$./*.nasm" ] && rm ./*.nasm #in case someone modifies the source code to produce .nasm files
+[ -f "$./*.fasm" ] && rm ./*.fasm #in case someone modifies the source code to produce .fasm files
 
 echo "Removing object files"
-rm ./*.o
-rm ./*.out #in case someone modifies the source code to produce .out files
-rm ./*.a #in case someone modifies the source code to produce .a files
+[ -f "$./*.o" ] && rm ./*.o
+[ -f "$./*.out" ] && rm ./*.out #in case someone modifies the source code to produce .out files
+[ -f "$./*.a" ] && rm ./*.a #in case someone modifies the source code to produce .a files
